+++ src/common/cassandra_client.py
@@ -2,10 +2,11 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-"""TODO."""
+"""Cassandra CQL client."""
 
 import logging
-from typing import List, Optional
+from contextlib import contextmanager
+from typing import Generator, List, Optional
 
 from cassandra.auth import PlainTextAuthProvider
 from cassandra.cluster import Cluster, Session
@@ -19,19 +20,18 @@ class CassandraClient:
     def __init__(
         self, hosts: List[str], user: Optional[str] = None, password: Optional[str] = None
     ):
+        self.auth_provider = (
+            PlainTextAuthProvider(username=user, password=password)
+            if user is not None and password is not None
+            else None
+        )
         self.hosts = hosts
-        self.user = user
-        self.password = password
-        self.auth_provider = None
-
-        if self.user is not None and self.password is not None:
-            self.auth_provider = PlainTextAuthProvider(username=self.user, password=self.password)
 
         return
 
     def create_keyspace(self, keyspace: str) -> None:
         """TODO."""
-        with _SessionContext(self, keyspace=None) as session:
+        with self._session() as session:
             query = (
                 """
                 CREATE KEYSPACE IF NOT EXISTS %s
@@ -44,23 +44,13 @@ class CassandraClient:
             )
             session.execute(query)
 
-
-class _SessionContext:
-    def __init__(self, client: CassandraClient, keyspace: Optional[str] = None):
-        self.client = client
-        self.keyspace = keyspace
-        self.cluster = None
-        self.session = None
-
-    def __enter__(self) -> Session:
-        self.cluster = Cluster(
-            contact_points=self.client.hosts, auth_provider=self.client.auth_provider
-        )
-        self.session = self.cluster.connect()
-        if self.keyspace:
-            self.session.set_keyspace(self.keyspace)
-        return self.session
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        if self.cluster:
-            self.cluster.shutdown()
+    @contextmanager
+    def _session(self, keyspace: str | None = None) -> Generator[Session]:
+        cluster = Cluster(contact_points=self.hosts, auth_provider=self.auth_provider)
+        session = cluster.connect()
+        if keyspace:
+            session.set_keyspace(keyspace)
+        try:
+            yield session
+        finally:
+            cluster.shutdown()
