+++ src/common/exceptions.py
@@ -2,7 +2,7 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-"""TODO."""
+"""Exceptions."""
 
 
 class HealthCheckFailedError(Exception):
